# swagger.yaml
openapi: 3.0.0
info:
    version: 1.0.0
    title: MatchUp
    description: BackEnd for the wbd project
...
paths:
    /products:
        get:
            summary: Get all the Products
            tags: [Product]
            responses:
                "200":
                    description: The list of all the Products
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Product"
        post:
            summary: Add a single Product
            tags: [Product]
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            $ref: "#/components/schemas/Product"
            responses:
                "200":
                    description: Product was created successfully
    /products/{productID}:
        get:
            summary: Get a product by id
            tags: [Product]
            parameters:
                - name: productID
                  in: path
                  required: true
                  description: Autogenerated Product ID by Mongoose
                  schema:
                      type: string
            responses:
                "200":
                    description: Successfully Fetched single product by the id.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Product"
        patch:
            summary: Update a Product by ID
            tags: [Product]
            parameters:
                - name: productID
                  in: path
                  required: true
                  description: Autogenerated Product ID by Mongoose
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/definitions/ProductPatchRequest"
            responses:
                "200":
                    description: Update a single product by the id.
                    content:
                        application/json:
                            schema: 
                                 $ref: "#/components/schemas/Product"
        delete:
            summary: Delete a product by id
            tags: [Product]
            parameters:
                - name: productID
                  in: path
                  required: true
                  description: Autogenerated Product ID by Mongoose
                  schema:
                      type: string
            responses:
                "200":
                    description: Successfully Deleted Product by the id.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Product"
    /blogs:
        get:
            summary: Get all the Blogs
            tags: [Blog]
            responses:
                "200":
                    description: The list of all the Blogs
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Blog"
        post:
            summary: Add a single Blog
            tags: [Blog]
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            $ref: "#/components/schemas/Blog"
            responses:
                "200":
                    description: Blog was created successfully
    /blogs/{blogID}:
        get:
            summary: Get a blog by id
            tags: [Blog]
            parameters:
                - name: blogID
                  in: path
                  required: true
                  description: Autogenerated Blog ID by Mongoose
                  schema:
                      type: string
            responses:
                "200":
                    description: Successfully Fetched single Blog by the id.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Blog"
        patch:
            summary: Update a Blog by ID
            tags: [Blog]
            parameters:
                - name: blogID
                  in: path
                  required: true
                  description: Autogenerated Blog ID by Mongoose
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/definitions/BlogPatchRequest"
            responses:
                "200":
                    description: Update a single Blog by the id.
                    content:
                        application/json:
                            schema: 
                                $ref: "#/components/schemas/Blog"
        delete:
            summary: Delete a blog by id
            tags: [Blog]
            parameters:
                - name: blogID
                  in: path
                  required: true
                  description: Autogenerated Blog ID by Mongoose
                  schema:
                      type: string
            responses:
                "200":
                    description: Successfully Deleted Blog by the id.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Blog"
...
components:
    schemas:
...
        Product:
            type: object
            required:
                - name
                - price
                - file
                - sellerUserId
            properties:
                name:
                    type: string
                    description: Name of the Product
                price:
                    type: number
                    description: Price of the Product that is being sell
                sellerUserId :
                    type: string
                    description: User Id of the Product Seller
                file:
                    type : file
                    description : Image of the Product
            example:
                name: Apple MacBook,
                price: 9999,
                cover : https://storage.googleapis.com/matchup-12345.appspot.com/macbook.jpg
                sellerUserId: 6246e608b1923e4b8f0f9ece
        Blog:
            type: object
            required:
                - title
                - description
                - file
                - sellerUserId
            properties:
                title:
                    type: string
                    description: Name of the Blog
                description:
                    type: string
                    description: Description about the Product
                sellerUserId :
                    type: string
                    description: User Id of the Blog Writer
                file:
                    type : file
                    description : Image of the Product
            example:
                title: Harry Porter,
                description: Fantacy,
                productImage: https://storage.googleapis.com/matchup-12345.appspot.com/24839.jpg,
                sellerUserId: 6246e608b1923e4b8f0f9ece
definitions:
...
  ProductPatchRequest:
    type: object
    items:
      $ref: "#/definitions/ProductPatchDocument"
  ProductPatchDocument: 
    description: A JSONPatch document for Products
    properties: 
     name: 
      type: string 
      description: Update Name of Product
     price: 
      type: number 
      description: Update Price of Product 
  BlogPatchRequest:
    type: object
    items:
      $ref: "#/definitions/BlogPatchDocument"
  BlogPatchDocument: 
    description: A JSONPatch document for Blogs
    properties: 
     title: 
      type: string 
      description: Update Name of Blog
     description: 
      type: string 
      description: Update Description of Blog
...